feat(frontend): enhance theme toggle with 0.5s View Transition animation
- Implemented native View Transitions API for the theme toggle, coordinated with React's `flushSync` for a seamless full-screen crossfade effect.
- Strictly enforced a 0.5-second animation duration per user requirements.
- Centralized the animation duration in a CSS variable (`--theme-transition-duration`) to maintain a single source of truth across CSS and JavaScript.
- Added accessibility support using Framer Motion's `useReducedMotion` hook to instantly toggle themes for users preferring reduced motion.
- Implemented robust fallback mechanisms for older browsers that dynamically parse the CSS duration variable, preventing timing mismatches.
