+++ Backend/package.json
@@ -23,7 +23,7 @@
     "express-validator": "7.3.2",
     "ioredis": "5.10.1",
     "jsonwebtoken": "9.0.3",
-    "mongodb": "7.1.1",
+    "mongodb": "7.1.0",
     "mongoose": "^9.4.1",
     "morgan": "1.10.1",
     "nodemailer": "^8.0.5",
